Rename step-progress to step-import
